hub/requirements.txt
PyYAML 5.3.1
paho-mqtt 1.5.0
python-keycloak-client 0.2.3
python-keycloak 0.25.0